requirements/developments.txt
asgiref==3.7.2
autopep8==2.0.4
chromedriver
Django==5.0
django-autoslug==1.9.9
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-factory_boy
django-filter==23.5
django-silk==5.0.4
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gprof2dot==2022.7.29
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
fuzzywuzzy
pycodestyle==2.11.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-Levenshtein
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
selenium
sqlparse==0.4.4
uritemplate==4.1.1
